Conflicts:
	gradle.properties
	modules/core/src/main/java/org/openlmis/core/domain/Right.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/GeographicZoneMapper.java
	modules/core/src/main/java/org/openlmis/core/repository/mapper/ProductMapper.java
	modules/core/src/main/java/org/openlmis/core/service/FacilityService.java
	modules/core/src/test/java/org/openlmis/core/service/RoleRightsServiceTest.java
	modules/openlmis-web/build.gradle
	modules/openlmis-web/src/main/webapp/WEB-INF/open-lmis-servlet.xml
	modules/openlmis-web/src/main/webapp/public/js/shared/services/services.js
	modules/openlmis-web/src/main/webapp/public/pages/navigation.html
	modules/requisition/src/test/resources/test-applicationContext-requisition.xml
	settings.gradle
	test-modules/functional-tests/src/test/java/org/openlmis/functional/testng-functional.xml

+++ modules/authentication/src/main/java/org/openlmis/authentication/UserToken.java
@@ -14,6 +14,6 @@ import lombok.Data;
 public class UserToken {
 
     private final String userName;
-    private final Integer userId;
+    private final Long userId;
     private final boolean authenticated;
 }
