+++ modules/core/src/main/java/org/openlmis/core/repository/mapper/RoleAssignmentMapper.java
@@ -26,40 +26,40 @@ public interface RoleAssignmentMapper {
   @Results(value = {@Result(property = "supervisoryNode", column = "supervisoryNodeId", javaType = SupervisoryNode.class,
       one = @One(select = "org.openlmis.core.repository.mapper.SupervisoryNodeMapper.getSupervisoryNode"))})
   List<RoleAssignment> getRoleAssignmentsWithGivenRightForAUser(@Param(value = "right") Right right,
-                                                                @Param(value = "userId") int userId);
+                                                                @Param(value = "userId") Long userId);
 
   @Insert("INSERT INTO role_assignments" +
       "(userId, roleId, programId, supervisoryNodeId) VALUES " +
       "(#{userId}, #{roleId}, #{programId}, #{supervisoryNodeId})")
-  int insertRoleAssignment(@Param(value = "userId") Integer userId,
-                           @Param(value = "programId") Integer programId, @Param(value = "supervisoryNodeId") Integer supervisoryNodeId, @Param(value = "roleId") Integer roleId);
+  int insertRoleAssignment(@Param(value = "userId") Long userId,
+                           @Param(value = "programId") Long programId, @Param(value = "supervisoryNodeId") Long supervisoryNodeId, @Param(value = "roleId") Long roleId);
 
   @Delete("DELETE FROM role_assignments WHERE userId=#{id}")
-  void deleteAllRoleAssignmentsForUser(int userId);
+  void deleteAllRoleAssignmentsForUser(Long userId);
 
   @Select("SELECT userId, programId, supervisoryNodeId, array_agg(roleId) as roleIdsAsString " +
       "FROM role_assignments " +
       "WHERE userId=#{userId} AND programId IS NOT NULL AND supervisoryNodeId IS NOT NULL " +
       "GROUP BY userId, programId, supervisoryNodeId ")
   @Results(value = {@Result(property = "supervisoryNode.id", column = "supervisoryNodeId")})
-  List<RoleAssignment> getSupervisorRoles(Integer userId);
+  List<RoleAssignment> getSupervisorRoles(Long userId);
 
 
   @Select("SELECT userId, programId, array_agg(roleId) as roleIdsAsString " +
       "FROM role_assignments " +
       "WHERE userId=#{userId} AND programId IS NOT NULL AND supervisoryNodeId IS NULL " +
       "GROUP BY userId, programId")
-  List<RoleAssignment> getHomeFacilityRoles(Integer userId);
+  List<RoleAssignment> getHomeFacilityRoles(Long userId);
 
   @Select("SELECT RA.userId, RA.programId, array_agg(RA.roleId) as roleIdsAsString " +
         "FROM role_assignments RA INNER JOIN role_rights RR ON RA.roleId = RR.roleId " +
         "WHERE RA.userId=#{userId} AND RA.programId=#{programId} AND RR.rightName = ANY (#{commaSeparatedRights}::VARCHAR[])  AND  supervisoryNodeId IS NULL " +
         "GROUP BY userId, programId")
 
-  List<RoleAssignment> getHomeFacilityRolesForUserOnGivenProgramWithRights(@Param("userId") Integer userId, @Param("programId") Integer programId,@Param("commaSeparatedRights") String commaSeparatedRights);
+  List<RoleAssignment> getHomeFacilityRolesForUserOnGivenProgramWithRights(@Param("userId") Long userId, @Param("programId") Long programId,@Param("commaSeparatedRights") String commaSeparatedRights);
 
 
   @Select({"SELECT RA.userId, array_agg(RA.roleId) as roleIdsAsString FROM role_assignments RA INNER JOIN roles R ON RA.roleId = R.id",
     "WHERE userId = #{userId} AND R.adminRole = true GROUP BY userId, supervisoryNodeId, programId"})
-  RoleAssignment getAdminRole(Integer userId);
+  RoleAssignment getAdminRole(Long userId);
 }
