+++ modules/core/src/test/java/org/openlmis/core/builder/ProgramProductBuilder.java
@@ -18,8 +18,8 @@ import static com.natpryce.makeiteasy.Property.newProperty;
 
 public class ProgramProductBuilder {
 
-  public static final Property<ProgramProduct, Integer> programId = newProperty();
-  public static final Property<ProgramProduct, Integer> productId = newProperty();
+  public static final Property<ProgramProduct, Long> programId = newProperty();
+  public static final Property<ProgramProduct, Long> productId = newProperty();
   public static final Property<ProgramProduct, Integer> dosagePerMonth = newProperty();
 
   private static Property<ProgramProduct, String> productCode = newProperty();
@@ -31,10 +31,10 @@ public class ProgramProductBuilder {
     @Override
     public ProgramProduct instantiate(PropertyLookup<ProgramProduct> lookup) {
       Product product = new Product();
-      product.setId(lookup.valueOf(productId, 1));
+      product.setId(lookup.valueOf(productId, 1L));
       product.setCode(lookup.valueOf(productCode, PRODUCT_CODE));
       Program program = new Program();
-      program.setId(lookup.valueOf(programId, 1));
+      program.setId(lookup.valueOf(programId, 1L));
       program.setCode(lookup.valueOf(programCode, PROGRAM_CODE));
       return new ProgramProduct(program, product, lookup.valueOf(dosagePerMonth, 1), true, new Money("100"));
     }
