+++ modules/shipment/src/main/java/org/openlmis/shipment/file/csv/handler/RawShipment.java
@@ -17,5 +17,5 @@ import org.openlmis.upload.annotation.ImportField;
 @NoArgsConstructor
 public class RawShipment implements Importable {
   @ImportField(type = "int",name = "Order Number")
-  private Integer orderNumber;
+  private Long orderNumber;
 }
